+++ src/app/components/table/demos/alternate.ts
@@ -1,21 +1,23 @@
+
+
+
 import { Component } from '@angular/core';
-import { NtColumnSortChange } from '@ng-tangram/components/table';
 
 @Component({
-  selector: 'demo-table-sort',
+  selector: 'demo-table-alternate',
   template: `
-  <nt-table #table [dataSource]="dataSource" (sortChange)="onSortChange($event)">
+  <nt-table [dataSource]="dataSource">
     <nt-column name="name">
       <nt-column-header *ntColumnHeaderDef>名称</nt-column-header>
       <nt-column-cell *ntColumnCellDef="let item">{{ item.name }}</nt-column-cell>
     </nt-column>
 
-    <nt-column name="age" align="center" sortable>
+    <nt-column name="age">
       <nt-column-header *ntColumnHeaderDef>年龄</nt-column-header>
       <nt-column-cell *ntColumnCellDef="let item">{{ item.age }}</nt-column-cell>
     </nt-column>
 
-    <nt-column name="address" align="right" sortable>
+    <nt-column name="address">
       <nt-column-header *ntColumnHeaderDef>地址</nt-column-header>
       <nt-column-cell *ntColumnCellDef="let item">{{ item.address }}</nt-column-cell>
     </nt-column>
@@ -23,13 +25,9 @@ import { NtColumnSortChange } from '@ng-tangram/components/table';
     <nt-header-row *ntHeaderRowDef="displayedColumns"></nt-header-row>
     <nt-row *ntRowDef="let row; columns: displayedColumns;"></nt-row>
   </nt-table>
-
-  <nt-callout>
-    {{ sortChange | json }}
-  </nt-callout>
   `
 })
-export class DemoTableSortComponent {
+export class DemoTableAlternateComponent {
 
   dataSource = [
     { id: 1, name: '张三', age: 20, address: '北京' },
@@ -39,10 +37,4 @@ export class DemoTableSortComponent {
   ];
 
   displayedColumns = ['name', 'age', 'address'];
-
-  sortChange: NtColumnSortChange;
-
-  onSortChange(change: NtColumnSortChange) {
-    this.sortChange = change;
-  }
 }